fix: fixing a error where an exception is not handled when a nonexistent item is removed

+++ app/services/cart.py
@@ -1,7 +1,7 @@
 from app.models import db, Cart
 from app.exceptions.itemDontExistsException import ItemDontExistsException
 from app.services.item import item_exists
-
+from app.exceptions.cartItemDontExistsException import cartItemDontExistsException
 
 def add_cart_item(user_id: int, item_id: int) -> None:
     if not item_exists(item_id):
@@ -16,6 +16,10 @@ def remove_cart_item(user_id: int, item_id: int) -> None:
         raise ItemDontExistsException("This item does not exists in our database.")
 
     cart_item_to_remove = Cart.query.filter_by(item_id=item_id, user_id=user_id).first()
+
+    if cart_item_to_remove is None:
+        raise cartItemDontExistsException("This cart item does not exists in our database.")
+
 
     db.session.delete(cart_item_to_remove)
     db.session.commit()
